Started the basic walking screen. Merged text eng.
Known Issues:

* Tile collision doesn't work perfectly. My guess is that it's a
tile/sprite size issue.

* I don't know how I feel about the text engine. It's snazzy and works
okay, but it's all based on lines in the single monolithic script file.
So to have an NPC talk, you say "Start at the script on line 10" or
whatever. The problem is, if you delete line 9, it means you have to go
through and change ALL of the NPCs to adjust for that. I don't love
that.