READ THESE COMMENTS
So I had to change the x origin again.

Here's why: I changed the playeravatar objects sprites to actual
sprites. If the x is not in the middle, it is not symmetrical and
therefore causes the sprite to jump like sixteen pixels if you turn
around.

So, the collision is fucked again. I knew your shit was too simple to be
true, but I didn't say anything because I was secretly pissed off about
how long I had worked on collisions.

Also, if you want to fix it, please fix it for objects as well as tiles,
or just forget about tiles because using the transparent wall objects
seems to be an easier method.