FUCKING fixed
But here's the rub:

If you check the room you can see how this is set up. The grid snap is
16 and everything is moved over 16 pixels to allow the x origin on the
sprites to be 16. I tried everything else I could think of, but this is
the only thing that worked.

In order to keep where you can and cannot place collision tiles
straight, there is an invisible blue grid you can copy and paste onto
the map. The objects self-destruct upon load.

The wall objects now make a tile at the collision layer the size of
themselves, so placing a wall is all you should need to do to make a
collision.

Oh, and I think there is a random collision tile floating around in the
middle of the grass. That's why the visible wall method is what I
prefer.